Updated code. Motors run Independantly with third motor activating when motors 1 and 2 are on
The previous version of the code was fine but we had a fault with the circuitry that was causing the vibrating motors to all drive at the same time. This problem has been resolved. The new code is a result of this error being fixed and the addition then of outputting a third motor when both motor 1 and 2 are active. This was done because the third ultrasound was faulty but while waiting for another, the GPIO pin was set to an output and is something useful at least. The motor is replaced with an LED acting as an indicator when both motors1 and 2 are active